use latest version of AerOpt and simplify QSettings paths

+++ Core/BoundaryPoint.cpp
@@ -38,8 +38,8 @@ uint BoundaryPoint::getSmoothing() {
     return mSmooth;
 }
 
-uint BoundaryPoint::getSmoothFactor() {
-    return mSmooth / 10;
+float BoundaryPoint::getSmoothFactor() {
+    return mSmooth / 10.0;
 }
 
 void BoundaryPoint::setSmoothing(uint smoothing) {
